Extract ComputedContingencyElement from DcSensitivityAnalysis
